Integrate code better with the playground for quick tests
Some more of the addressing mode code implemented

+++ MOS6502/CPU/CPU.swift
@@ -44,7 +44,7 @@ class CPU {
     
     var memory: Memory
     var calculationSum: Int = 0
-    var memoryFetchedValue: UInt8 = 0
+    var memoryFetchedValue: UInt16 = 0
     
     init(with memory: Memory) {
         self.memory = memory
